Create one function for setting up a ParsingPage with the appropriate variable context.
